# api/openapi.yaml
openapi: 3.1.0
info:
  title: CoachPulse API
  version: "1.0.0"
  description: |
    The API behind CoachPulse — a business operating system for independent
    personal trainers.

    Two conventions run through everything here and are worth reading before
    the endpoints.

    **Money is integer minor units.** Every amount is a whole number of cents,
    fils or agorot, paired with an ISO-4217 code. There are no floats. A
    ledger that rounds is a ledger that cannot be reconciled.

    **Errors carry a stable machine code.** Clients branch on `error.code`,
    never on `error.message`: the message is for humans and may be reworded,
    the code is part of the contract. `insufficient_credits` carries the
    remaining balance in `meta`, which is what lets the app offer a renewal
    at the moment a pack runs out.

    Measurements follow money's lead: loads are grams, RPE is tenths (85 is
    RPE 8.5), body fat is basis points. Display units are the client's choice;
    the stored value never rounds.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.coachpulse.io
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/signup:
    post:
      tags: [Authentication]
      summary: Register a trainer and their business
      description: Provisions the tenant, its owner, its chart of accounts and
        its exercise library in one transaction.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupRequest" }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "409": { $ref: "#/components/responses/Conflict" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/auth/login:
    post:
      tags: [Authentication]
      summary: Sign in
      security: []
      parameters:
        - $ref: "#/components/parameters/RefreshTransport"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401":
          description: |
            Wrong password and unknown account are deliberately
            indistinguishable, so this endpoint cannot be used to discover
            which emails are registered.

            With two-step sign-in on, a right password answers
            `mfa_required` with `meta.mfa_token`: exchange it, with the code
            from the authenticator app, at `/v1/auth/mfa`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Exchange a refresh token for a new pair
      description: |
        Refresh tokens rotate on every use and are tracked as a family.
        Presenting one that has already been rotated means two parties hold
        it, so the entire family is revoked and the response is
        `refresh_token_reused`.

        The web build sends an empty body and `X-Refresh-Transport: cookie`;
        the token comes from its httpOnly cookie instead.
      security: []
      parameters:
        - $ref: "#/components/parameters/RefreshTransport"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Rotated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/auth/mfa:
    post:
      tags: [Authentication]
      summary: Finish a two-step sign-in
      description: The code is either the current one from the authenticator
        app or one of the single-use recovery codes. A code already used is
        refused even inside its thirty seconds.
      security: []
      parameters:
        - $ref: "#/components/parameters/RefreshTransport"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mfa_token, code]
              properties:
                mfa_token: { type: string, description: From the login's `mfa_required` meta }
                code: { type: string, examples: ["492039", "7KQ2M-X9PDA"] }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/auth/password/forgot:
    post:
      tags: [Authentication]
      summary: Email a password-reset link
      description: Accepted whether or not the address has an account, so the
        endpoint cannot be used to find out who does. The link works once,
        for an hour.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202": { description: Sent if the address is registered }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/auth/password/reset:
    post:
      tags: [Authentication]
      summary: Set a new password from a reset link
      description: Signs out every device, since whoever asked for the reset
        may not be the only one who knew the old password.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 12 }
      responses:
        "204": { description: Password changed }
        "400":
          description: A weak password, or `reset_link_invalid` for a link
            that has expired or been used.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RefreshTransport:
      name: X-Refresh-Transport
      in: header
      schema: { type: string, enum: [cookie] }
      description: |
        `cookie` puts the refresh token in an httpOnly, SameSite=Strict
        cookie instead of the body, for the web build — a browser has nowhere
        safe to keep one. Refresh and logout then read it from the cookie.
...
  responses:
...
    TooManyRequests:
      description: "`rate_limited`: too many attempts; wait and try again."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    BadRequest:
      description: Malformed or failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Conflict:
      description: A state precondition was violated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Stable and part of the contract — branch on this, not on the
                message.
              examples: [insufficient_credits, overpayment, scheduling_conflict,
                         invoice_not_payable, invalid_state_transition,
                         idempotency_key_reused, validation_failed]
            message: { type: string, description: For humans; may be reworded }
            fields:
              type: object
              additionalProperties: { type: string }
              description: Per-field validation detail
            meta:
              type: object
              additionalProperties: true
              description: |
                Structured context the client can act on. An
                `insufficient_credits` refusal carries `remaining` and
                `required`, which is what drives the renewal prompt.
            request_id: { type: string }
    Session:
      type: object
      properties:
        account:
          type: object
          properties:
            user_id: { type: string, format: uuid }
            tenant_id: { type: string, format: uuid }
            email: { type: string }
            display_name: { type: string }
            role: { type: string }
            currency: { type: string }
        tokens:
          type: object
          properties:
            access_token: { type: string }
            refresh_token: { type: string }
            expires_at: { type: string, format: date-time }
            token_type: { type: string, examples: [Bearer] }
            refresh_expires_at:
              type: string
              format: date-time
              description: When this device is signed out if unused — the
                practice's session timeout from now.
...
    SignupRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12 }
        display_name: { type: string }
        business_name: { type: string }
        currency: { type: string, default: USD }
        timezone: { type: string, examples: [Europe/Berlin] }
...